Warrenn Enslin

+++ contracts/Example3.sol
@@ -1,10 +1,10 @@
 pragma solidity ^0.8.4;
 
 contract Example3 {
-    function getOwnerMargin(uint8 amount)
+    function getOwnerMargin(uint256 amount)
         external
         pure
-        returns (uint8 amountForSender, uint8 amountForOwner)
+        returns (uint256 amountForSender, uint256 amountForOwner)
     {
         amountForSender = (amount * 70) / 100;
         amountForOwner = (amount * 30) / 100;
